- Added  enable timer output - Updated contributing markdown - Removed defined checks against __cplusplus - Tinkered around with virtual memory management - Reordered some headers - Introduced new type paddr_t and vaddr_t - Changed uintptr_t to paddr_t where necessary - Changed void pointer to new type vaddr_t - Added timer initialization to startup for bcm 2709 and 2710 - Extended mailbox property function by bunch of tags - Changed serial port initialization to gather clock rate via mailbox with dynamic calculation of ibrd and fbrd - Cleaned up timer code a bit and encapsulated code for bcm 2709 and 2710 within pre processor checks

+++ src/lib/stdc/stdio/printf.c
@@ -138,7 +138,7 @@ int printf( const char* restrict format, ... ) {
       case 'u':
       case 'p':
         if ( 'p' == *format ) {
-          un = ( uintptr_t )va_arg( parameters, void* );
+          un = ( uint32_t )va_arg( parameters, void* );
           radix = 16;
         } else {
           un = va_arg( parameters, uint32_t );
